+++ CHANGELOG.md
@@ -11,7 +11,7 @@ In order to build the DUNE core modules you need at least the following software
 - The construction of `FiedlMatrix` and `FieldVector` from `std::initializer_list`
   is now `constexpr`.
 - C++ compilers LLVM Clang >= 10 or GCC g++ >= 9
-- CMake >= 3.13
+- CMake >= 3.16
 - Optional: pkg-config to find other optional dependencies
 - Optional: Python >= 3.7 for Python bindings
 
@@ -94,6 +94,10 @@ In order to build the DUNE core modules you need at least the following software
 
 - The cmake options `CXX_MAX_STANDARD`, `CXX_MAX_SUPPORTED_STANDARD` and `DISABLE_CXX_VERSION_CHECK`
   are removed. The cmake function `dune_require_cxx_standard()` is now deprecated.
+
+- Deprecate CMake macro `message_verbose`. Use `message(VERBOSE "message text")` instead.
+  This macro will be removed after Dune 2.10.
+
 
 ## Deprecations and removals